SettingsPlotX classes into separate files
