+++ Mappalachia/Mappalachia.csproj
@@ -76,7 +76,10 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Class\SettingsPlotStyle.cs" />
     <Compile Include="Class\SettingsFileExport.cs" />
+    <Compile Include="Class\SettingsPlotHeatmap.cs" />
+    <Compile Include="Class\SettingsPlotTopograph.cs" />
     <Compile Include="Class\Space.cs" />
     <Compile Include="Class\SettingsSpace.cs" />
     <Compile Include="Class\HeatMapGridSquare.cs" />
@@ -209,6 +212,7 @@
       <Version>5.0.0</Version>
     </PackageReference>
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <PropertyGroup>
     <PostBuildEvent>xcopy $(ProjectDir)\data\* $(TargetDir)\data\ /y
